+++ llama_index/response_synthesizers/no_text.py
@@ -2,6 +2,7 @@ from typing import Any, Sequence
 from llama_index.response_synthesizers.base import BaseSynthesizer
 from llama_index.types import RESPONSE_TEXT_TYPE
 
+
 class NoText(BaseSynthesizer):
     def get_response(
         self,
@@ -10,7 +11,7 @@ class NoText(BaseSynthesizer):
         **response_kwargs: Any,
     ) -> RESPONSE_TEXT_TYPE:
         return ""
-    
+
     async def aget_response(
         self,
         query_str: str,
@@ -18,4 +19,3 @@ class NoText(BaseSynthesizer):
         **response_kwargs: Any,
     ) -> RESPONSE_TEXT_TYPE:
         return ""
-